Use script instead of bazel for elafros/elafros.
For the prow test jobs for the elafros/elafros repo, run
tests/presubmit-tests.sh instead of "bazel test ...".

Using a shell script for running the presubmit tests
gives us more flexibility, allowing them to be multi-step
and have complex behavior.

Furthermore, it puts the entire test config in a single place
(the repo) instead of two places (the repo and the prow config file).

Also remove the word "bazel" from the job name to reflect that it's
not simply a bazel job anymore.

Change-Id: Ia11f8178bc769e0ebf228e07ad7487cdd1894179