explicitly supply object name for GCS uploading
The script triggered by the postsubmit prow job (for test coverage) is
using $PULL_PULL_SHA as object name. However, that var is only available
in pull request. Postsubmit is not a pull request and thus doesn't have
the var. We are not supplying $PULL_BASE_SHA as the explicit file name
(the first optional argument of the sh script) as a solution.

Change-Id: I7cd28853c9aa85ce6ec86c5ebfdff35f839db1e5