Make elafros test jobs use a different GCS bucket
The ela-prow GCS bucket will be used for test logs. This bucket is viewable by
members of the Elafros project team members outside Google.

The default gke-prow GCS bucket is anot accessible by such members, so they
can't see the logs for the latest test invocation for their PRs. This is simpler
than bringing up a new gubernator and setting up ACLs with Cloud IAP. The raw
log can be fetched by using Pantheon or by assembling the URL to the log based
on the information added to the PR thread by prow.

Change-Id: I800be180e35bdd75b3f9f361d345e0e0bf7fcef9