Create a periodic CI job for Elafros
Tests only run during presubmit, so it's hard to measure how stable they are
(e.g., they might fail because of a real bug in the PR). Running the tests on
a regular basis (e.g., every 2h) against the master branch will not only give
a measure of how stable the tests are, but also if the current code at head is
broken.

Change-Id: Ibb4f987011a4c7ba8359d0a1147cc3b82b5c15ae