nomos-ci-gke-110: adds this new job
The job runs CI type tests against a GKE 1.10.x cluster.
We will be adding a few more of similar tests to handle various
other nomos configs of interest.

The cluster to use is set up out of band and is selected based on the
name of the job: USER named "foo" will use "foo-cluster-1".  This
approach works well for our current use cases, so we'd just use it for
the time being.

Bug: 113165924
Change-Id: I56c723e753fc8c806f89a7dd1920fe9315353231