Setup 2 configs for Batch repository Config 1: pull-batch-presubmit         This is run on every presubmit. Config 2: ci-bigcomputeeng-batch-acceptance-tests         This is run every 1 hour.
Other Miscellaneous changes:
Removing GKE Config as we are not yet ready to run on GKE yet.
Removing GCE Config as that is not working yet and we need special
version of k8s to set this up. Will be checked back in once we have the
special version of k8s.

Change-Id: Ie36a9bc29e2077eebcce3cff2223d7b77d22c3b8