nomos-release: run the job once in 24 hours
We only need to build a release once a day, should be enough
to get at least one release candidate per week.

The old setting ran the job every 1 minute during development.
Now that we had a few successful runs (see go/nomos-cd), it
is time to roll back to the original once-a-day plan.

Bug: 114327420
Change-Id: I966761df1c47d2f37e40a9fe45efb869f594ca68