mount GCP key as volume in CNRM test jobs
As CNRM exposes a Kubernetes API surface for existing GCP services, our
integration tests require our Prow test jobs to reach out to GCP and
mutate other GCP services (like Cloud Spanner, Cloud Pub/Sub, etc.). For
this, we need to inject a GCP service account with the appropriate
permissions for our jobs to act as.

Erick Fejta helped us inject a GCP SA key into a secret in the
'test-pods' namespace in the internal Prow cluster. This commit mounts
this as a volume in our test pods.

Change-Id: I79cf79cf63f0888a92237b628ecf587209144c7d