Revert nomos-operator-release kubekins upgrade
https://gke-internal-review.googlesource.com/c/test-infra/+/35260
upgraded the version of kubekins that the nomos-operator-release job was
using, but the subsequent run failed earlier than the run previous.

Certainly, we should upgrade this at some point, but for now I would
prefer to get the job debugged, and then try upgrading.

This change reverts the change linked above.

Change-Id: I1841a21cf8f0451b1f533c2f7f5eb03f86c983bf