Upgrade nomos-operator-release-staging kubekins version
The nomos-operator-release-staging job uses `go mod`, which does not
exist in the 20180730 version this job is currently using. This recent
version of kubekins does have support for go mod.

Note that a kubekins version upgrade was previously reverted but we
determined afterwards that the upgrade was not the cause of the observed
problem

Change-Id: Ibac5e9f2fbccb19fd600da86444ca555693d497a