nomos-operator set kubekins version
Previously, we upgraded to the most recent version of kubekins, which
was built on 1/08/2019. We needed >= 1.11 because nomos-operator uses go
mod, which was included in 1.11.
However, due to an issue with go mod detailed at:
https://github.com/hashicorp/go-discover/issues/68
there is a checksum mismatch during build that causes the build of the
nomos operator to fail.

Now, one possible option would be to upgrade the go version of all devs
on this repo past go1.11.4, but from what I can tell that isn't easily
available in glinux packages yet, and so feels rather frictiony and
unpleasant. So, the alternate approach is to use a different version of
kubekins that is recent enough to support go mod (>=1.11) but not so
recent as to have the change that causes the checksum mismatch
(<1.11.4). I found one such version from early december.

As 1.11.4 becomes easily available to devs, we will need to update
go.sum and the kubekins container for this job. This is a
minimum-friction holdover until then.

Change-Id: I58075d596fc2d495cad9bf5af35d0b2c3ae36240